Move Component class to component package
